+++ app/src/main/java/com/belcobtm/data/rest/transaction/request/SellRequest.kt
@@ -5,6 +5,6 @@ data class SellRequest(
     val type: Int,
     val cryptoAmount: Double,
     val price: Double,
-    val fiatAmount: Double,
+    val fiatAmount: Int,
     val serviceFee: Double
 )
